#2 and #3 fixed.

+++ app.js
@@ -93,24 +93,24 @@ app.get( '/api/source/:source_id', async function( req, res ){
   res.end();
 });
 
-app.get( '/api/jsons/:source_id', async function( req, res ){
+app.get( '/api/strings/:source_id', async function( req, res ){
   res.contentType( 'application/json; charaset=utf-8' );
 
   var _checked = ( 'checked' in req.query ? req.query.checked : null );
   var checked = ( _checked ? parseInt( _checked ) : -1 );
   var source_id = req.params.source_id;
-  var jsons = await app.readJsonsBySourceId( source_id, checked );
-  res.write( JSON.stringify( { status: true, source_id: source_id, jsons: jsons }, null, 2 ) );
+  var strings = await app.readStringsBySourceId( source_id, checked );
+  res.write( JSON.stringify( { status: true, source_id: source_id, strings: strings }, null, 2 ) );
   res.end();
 });
 
-app.post( '/api/json/:json_id/:checked', async function( req, res ){
+app.post( '/api/string/:string_id/:checked', async function( req, res ){
   res.contentType( 'application/json; charaset=utf-8' );
 
-  var json_id = req.params.json_id;
+  var string_id = req.params.string_id;
   var checked = parseInt( req.params.checked );
-  var b = await app.setJsonChecked( json_id, checked );
-  res.write( JSON.stringify( { status: true, json_id: json_id, checked: checked, b: b }, null, 2 ) );
+  var b = await app.setStringChecked( string_id, checked );
+  res.write( JSON.stringify( { status: true, string_id: string_id, checked: checked, b: b }, null, 2 ) );
   res.end();
 });
 
@@ -118,7 +118,7 @@ app.delete( '/api/url/:url_id', async function( req, res ){
   res.contentType( 'application/json; charaset=utf-8' );
 
   var url_id = req.params.url_id;
-  var r = await app.removeDataByUrlId( url_id );
+  var r = await app.removeDataByUrlId( url_id, true );
   res.write( JSON.stringify( { status: true, result: r }, null, 2 ) );
   res.end();
 });
@@ -245,47 +245,14 @@ app.readSourcesByUrlId = async function( url_id ){
   });
 }
 
-app.readSourceBodyById = async function( source_id ){
+app.readStringsBySourceId = async function( source_id, checked ){
   return new Promise( async function( resolve, reject ){
     var conn = null;
-    var body = '';
+    var strings = [];
     try{
       conn = await pg.connect();
       if( conn ){
-        var sql = "select body from sources where id = $1";
-        var query = { text: sql, values: [ source_id ] };
-        conn.query( query, function( err, result ){
-          if( err ){
-            console.log( err );
-          }else{
-            if( result.rows.length > 0 && result.rows[0].source ){
-              body = result.rows[0].body;
-            }
-          }
-          resolve( body );
-        });
-      }else{
-        resolve( body );
-      }
-    }catch( e ){
-      console.log( e );
-      resolve( body );
-    }finally{
-      if( conn ){
-        conn.release();
-      }
-    }
-  });
-}
-
-app.readJsonsBySourceId = async function( source_id, checked ){
-  return new Promise( async function( resolve, reject ){
-    var conn = null;
-    var jsons = [];
-    try{
-      conn = await pg.connect();
-      if( conn ){
-        var sql = "select id, start_index, end_index, body, checked, created from jsons where source_id = $1 order by created";
+        var sql = "select id, string, checked, created from strings where source_id = $1 order by created";
         var query = { text: sql, values: [ source_id ] };
         if( checked == 1 || checked == 0 ){
           query.text += ' and checked = $2';
@@ -298,21 +265,21 @@ app.readJsonsBySourceId = async function( source_id, checked ){
             if( result.rows.length > 0 && result.rows[0].id ){
               try{
                 for( var i = 0; i < result.rows.length; i ++ ){
-                  jsons.push( result.rows[i] );
+                  strings.push( result.rows[i] );
                 }
               }catch( e ){
                 console.log( e );
               }
             }
           }
-          resolve( jsons );
+          resolve( strings );
         });
       }else{
-        resolve( jsons );
+        resolve( strings );
       }
     }catch( e ){
       console.log( e );
-      resolve( jsons );
+      resolve( strings );
     }finally{
       if( conn ){
         conn.release();
@@ -321,15 +288,15 @@ app.readJsonsBySourceId = async function( source_id, checked ){
   });
 }
 
-app.setJsonChecked = async function( json_id, checked ){
+app.setStringChecked = async function( string_id, checked ){
   return new Promise( async function( resolve, reject ){
     var b = false;
     var conn = null;
     try{
       conn = await pg.connect();
       if( conn ){
-        var sql = "update jsons set checked = $1 where id = $2";
-        var query = { text: sql, values: [ checked, json_id ] };
+        var sql = "update strings set checked = $1 where id = $2";
+        var query = { text: sql, values: [ checked, string_id ] };
         conn.query( query, function( err, result ){
           if( err ){
             console.log( err );
@@ -384,7 +351,36 @@ app.storeUrl = async function( url ){
   });
 }
 
-app.storeSource = async function( url_id, source, body ){
+app.updateUrl = async function( url_id ){
+  return new Promise( async function( resolve, reject ){
+    var conn = null;
+    try{
+      conn = await pg.connect();
+      if( conn ){
+        var ts = ( new Date() ).getTime();
+        var sql = "update urls set created = $1 where id = $2";
+        var query = { text: sql, values: [ ts, url_id ] };
+        conn.query( query, function( err, result ){
+          if( err ){
+            //console.log( err );
+          }
+          resolve( { status: true, id: url_id } );
+        });
+      }else{
+        resolve( { status: false, error: 'no conn' } );
+      }
+    }catch( e ){
+      console.log( e );
+      resolve( { status: false, error: e } );
+    }finally{
+      if( conn ){
+        conn.release();
+      }
+    }
+  });
+}
+
+app.storeSource = async function( url_id, source ){
   return new Promise( async function( resolve, reject ){
     var conn = null;
     try{
@@ -392,8 +388,39 @@ app.storeSource = async function( url_id, source, body ){
       if( conn ){
         var id = crypto.randomUUID();
         var ts = ( new Date() ).getTime();
-        var sql = "insert into sources( id, url_id, source, body, created ) values( $1, $2, $3, $4, $5 )";
-        var query = { text: sql, values: [ id, url_id, source, body, ts ] };
+        var sql = "insert into sources( id, url_id, source, created ) values( $1, $2, $3, $4 )";
+        var query = { text: sql, values: [ id, url_id, source, ts ] };
+        conn.query( query, function( err, result ){
+          if( err ){
+            //console.log( err );
+          }
+          console.log( err, result );
+          resolve( { status: true, id: id } );
+        });
+      }else{
+        resolve( { status: false, error: 'no conn' } );
+      }
+    }catch( e ){
+      console.log( e );
+      resolve( { status: false, error: e } );
+    }finally{
+      if( conn ){
+        conn.release();
+      }
+    }
+  });
+}
+
+app.storeString = async function( source_id, string ){
+  return new Promise( async function( resolve, reject ){
+    var conn = null;
+    try{
+      conn = await pg.connect();
+      if( conn ){
+        var id = crypto.randomUUID();
+        var ts = ( new Date() ).getTime();
+        var sql = "insert into strings( id, source_id, string, created ) values( $1, $2, $3, $4 )";
+        var query = { text: sql, values: [ id, source_id, string, ts ] };
         conn.query( query, function( err, result ){
           if( err ){
             //console.log( err );
@@ -414,34 +441,25 @@ app.storeSource = async function( url_id, source, body ){
   });
 }
 
-app.storeJson = async function( source_id, start_index, end_index, body ){
-  return new Promise( async function( resolve, reject ){
-    var conn = null;
-    try{
-      conn = await pg.connect();
-      if( conn ){
-        var id = crypto.randomUUID();
-        var ts = ( new Date() ).getTime();
-        var sql = "insert into jsons( id, source_id, start_index, end_index, body, created ) values( $1, $2, $3, $4, $5, $6 )";
-        var query = { text: sql, values: [ id, source_id, start_index, end_index, body, ts ] };
-        conn.query( query, function( err, result ){
-          if( err ){
-            //console.log( err );
-          }
-          resolve( { status: true, id: id } );
-        });
-      }else{
-        resolve( { status: false, error: 'no conn' } );
-      }
-    }catch( e ){
-      console.log( e );
-      resolve( { status: false, error: e } );
-    }finally{
-      if( conn ){
-        conn.release();
-      }
+// テキストから JSON 表現を全て抜き出して配列で返す
+app.detectTargets = function( text ){
+  var targets = [];
+
+  //var regex = /[A-Za-z0-9.#@!%&|_-]{12,}/g;
+  var regex = /[A-Za-z0-9.#@!%_-]{12,}/g;
+  var results = text.match( regex );
+  //console.log( results );
+  for( var i = 0; i < results.length; i ++ ){
+    var regexnum = /[0-9]/;
+    var regexalphabet = /[A-Za-z]/;
+    var r1 = results[i].match( regexnum );
+    var r2 = results[i].match( regexalphabet );
+    if( r1 && r1.length && r2 && r2.length ){
+      targets.push( results[i] );
     }
-  });
+  }
+
+  return targets;
 }
 
 // テキストから JSON 表現を全て抜き出して配列で返す
@@ -525,13 +543,13 @@ app.JSONdetect = function( texts ){
   return jsons;
 }
 
-app.deleteJsonsBySourceId = async function( source_id ){
+app.deleteStringsBySourceId = async function( source_id ){
   return new Promise( async function( resolve, reject ){
     var conn = null;
     try{
       conn = await pg.connect();
 
-      var sql = "delete from jsons where source_id = $1";
+      var sql = "delete from strings where source_id = $1";
       var query = { text: sql, values: [ source_id ] };
       conn.query( query, function( err, result ){
         if( err ){
@@ -562,8 +580,67 @@ app.crawlUrl = async function( url ){
         if( body0 ){
           app.existUrl( url ).then( function( url_id ){
             if( url_id ){
-              console.log( 'already existed url: ' + url );
-              resolve( false );
+              app.updateUrl( url_id ).then( function( r0 ){
+                if( r0 && r0.status ){
+                  //console.log( { url }, { r0 } );
+                  $('script').each( async function(){
+                    var src = $(this).attr( 'src' );
+                    if( src && src.toLowerCase().endsWith( '.js' ) ){
+                      var src1 = src;
+                      if( !src.startsWith( 'http' ) ){
+                        //. "<script src='http(s)://.....'> という指定になってなかった場合
+                        if( src.startsWith( '//' ) ){
+                          //. "<script src='//.....'> という指定だった場合（ホスト名あり）
+                          if( url.startsWith( 'https' ) ){
+                            src1 = 'https:' + src;
+                          }else{
+                            src1 = 'http:' + src;
+                          }
+                        }else if( src.startsWith( '/' ) ){
+                          //. "<script src='/.....'> という指定だった場合（ホスト名なしのフルパス指定）
+                          var tmp = url.split( '/' );  //. [ 'https:' , '', 'www.example.com', 'path1', 'hello.html' ]
+                          while( tmp.length > 3 ){
+                            //. 元の URL から "https://www.example.com" 部分を取り出して補完する
+                            tmp.pop();
+                          }
+                          src1 = tmp.join( '/' ) + src;
+                        }else{
+                          //. "<script src='.....'> という指定だった場合（ホスト名なしの相対パス指定）
+                          if( url.endsWith( '/' ) ){
+                            src1 = url + src;
+                          }else{
+                            src1 = url + '/' + src;
+                          }
+                        }
+                      }
+  
+                      console.log( src1 + ' : ' );  //. "http(s)://....." という形に補完された JavaScript URL
+                      var r1 = await client.fetch( src1, {}, 'utf-8' );
+                      if( r1 && r1.body ){
+                        try{
+                          var body1 = r1.body;
+                          var r2 = await app.storeSource( url_id, src1 );
+                          if( r2 && r2.status ){
+                            //console.log( { src1 }, { r2 } );
+                            var source_id = r2.id;
+                            await app.deleteStringsBySourceId( source_id );
+            
+                            var results = app.detectTargets( body1 );
+                            //console.log( {results} );
+                            for( var i = 0; i < results.length; i ++ ){
+                              var r3 = await app.storeString( source_id, results[i] );
+                              console.log( i, results[i] )
+                            }
+                          }
+                        }catch( e ){
+                          console.log( {e} );
+                        }
+                      }
+                    }
+                    resolve( true );
+                  });
+                }
+              });
             }else{
               app.storeUrl( url ).then( function( r0 ){
                 if( r0 && r0.status ){
@@ -605,17 +682,17 @@ app.crawlUrl = async function( url ){
                       if( r1 && r1.body ){
                         try{
                           var body1 = r1.body;
-                          var r2 = await app.storeSource( url_id, src1, body1 );
+                          var r2 = await app.storeSource( url_id, src1 );
                           if( r2 && r2.status ){
                             //console.log( { src1 }, { r2 } );
                             var source_id = r2.id;
-                            await app.deleteJsonsBySourceId( source_id );
+                            await app.deleteStringsBySourceId( source_id );
             
-                            var results = app.JSONdetect( body1 );
+                            var results = app.detectTargets( body1 );
                             //console.log( {results} );
                             for( var i = 0; i < results.length; i ++ ){
-                              var r3 = await app.storeJson( source_id, results[i].start_index, results[i].end_index, results[i].text );
-                              console.log( i, results[i].text.substr( 0, 40 ) )
+                              var r3 = await app.storeString( source_id, results[i] );
+                              console.log( i, results[i] )
                             }
                           }
                         }catch( e ){
@@ -641,7 +718,7 @@ app.crawlUrl = async function( url ){
   });
 }
 
-app.removeDataByUrlId = async function( url_id ){
+app.removeSourcesAndStringsByUrlId = async function( url_id ){
   return new Promise( async function( resolve, reject ){
     var conn = null;
     try{
@@ -652,7 +729,7 @@ app.removeDataByUrlId = async function( url_id ){
         conn.query( query0, function( err0, result0 ){
           if( err0 ){
             console.log( err0 );
-            resolve( [ false, false, false ] );
+            resolve( [ false, false ] );
           }else{
             //console.log( {result0} );
             if( result0.rows.length > 0 && result0.rows[0].id ){
@@ -666,7 +743,7 @@ app.removeDataByUrlId = async function( url_id ){
               }
 
               if( source_ids.length > 0 ){
-                var sql1 = "delete from jsons where source_id = any($1::varchar[])";
+                var sql1 = "delete from strings where source_id = any($1::varchar[])";
                 var query1 = { text: sql1, values: [ source_ids ] };
                 conn.query( query1, function( err1, result1 ){
                   if( err1 ){
@@ -680,44 +757,60 @@ app.removeDataByUrlId = async function( url_id ){
                       console.log( err2 );
                     }
 
-                    var sql3 = "delete from urls where id = $1";
-                    var query3 = { text: sql3, values: [ url_id ] };
-                    conn.query( query3, function( err3, result3 ){
-                      if( err3 ){
-                        console.log( err3 );
-                      }
-                      resolve( [ true, true, true ] );
-                    });
+                    resolve( [ true, true ] );
                   });
                 });
               }else{
-                var sql3 = "delete from urls where id = $1";
-                var query3 = { text: sql3, values: [ url_id ] };
-                conn.query( query3, function( err3, result3 ){
-                  if( err3 ){
-                    console.log( err3 );
-                  }
-                  resolve( [ true, false, false ] );
-                });
+                resolve( [ false, false ] );
               }
             }else{
-              var sql3 = "delete from urls where id = $1";
-              var query3 = { text: sql3, values: [ url_id ] };
-              conn.query( query3, function( err3, result3 ){
-                if( err3 ){
-                  console.log( err3 );
-                }
-                resolve( [ true, false, false ] );
-              });
+              resolve( [ false, false ] );
             }
           }
         });
       }else{
-        resolve( [ false, false, false ] );
+        resolve( [ false, false ] );
       }
     }catch( e ){
       console.log( e );
-      resolve( [ false, false, false ] );
+      resolve( [ false, false ] );
+    }finally{
+      if( conn ){
+        conn.release();
+      }
+    }
+  });
+}
+
+app.removeDataByUrlId = async function( url_id, delete_data ){
+  return new Promise( async function( resolve, reject ){
+    var r = [ null, null ];
+    if( delete_data ){
+      r = await app.removeSourcesAndStringsByUrlId( url_id );
+    }
+
+    try{
+      conn = await pg.connect();
+      if( conn ){
+        var sql3 = "delete from urls where id = $1";
+        var query3 = { text: sql3, values: [ url_id ] };
+        conn.query( query3, function( err3, result3 ){
+          if( err3 ){
+            console.log( err3 );
+            r.unshift( false );
+          }else{
+            r.unshift( true );
+          }
+          resolve( r );
+        });
+      }else{
+        r.unshift( false );
+        resolve( r );
+      }
+    }catch( e ){
+      console.log( e );
+      r.unshift( false );
+      resolve( r );
     }finally{
       if( conn ){
         conn.release();
